feat: change height and weight to Int in Profile

+++ app/src/main/java/com/example/myfitnessapp/model/Profile.kt
@@ -1,6 +1,6 @@
 package com.example.myfitnessapp.model
 
 data class Profile(
-    val height: Double,
-    val weight: Double
+    val height: Int,
+    val weight: Int
 )
